Create new jenkins pipeline
This should build and push a docker image that has a tag which matches the API version. The new image is then supplied to the kubernetes deployment.

There a some conceptual problems with this, like the inability to undo the rollout in the event of failure, but I just want to make sure this works for now.